feat: sns neuron modals
